% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aichi13_indicator.R
\name{Aichi13_indicator}
\alias{Aichi13_indicator}
\title{Using Ex-situ and In-situ gap analysis results to calculate Aichi13 indicator value}
\usage{
Aichi13_indicator(Species_list, FCSc_mean_df)
}
\arguments{
\item{Species_list}{A species list to calculate metrics.}

\item{FCSc_mean_df}{A data frame object result of the function FCSc_mean
Species_list,FCSc_mean_df}
}
\value{
this function returns a data frame object with the following columns:

\tabular{lcc}{
 opt \tab \cr
 N_HP \tab \cr
 N_MP \tab \cr
 N_LP \tab \cr
 N_SC \tab \cr
 N_LP_SC  \tab \cr
 P_HP \tab \cr
 P_MP \tab \cr
 P_LP \tab \cr
 P_SC \tab \cr
 P_LP_SC \tab \cr
}
}
\description{
This function uses a dataframe product of use the function FCSc_mean and
 provides a comprehensive summary including a conservation indicator used in Aichi13 documentation
}
\examples{
##Obtaining occurrences from example
data(CucurbitaData)
##Obtaining species names from the data
Cucurbita_splist <- unique(CucurbitaData$taxon)
##Obtaining raster_list
data(CucurbitaRasters)
CucurbitaRasters <- raster::unstack(CucurbitaRasters)
##Obtaining protected areas raster
data(ProtectedAreas)
##Obtaining ecoregions shapefile
data(ecoregions)

#Running all three Ex-situ gap analysis steps using FCSex function
FCSex_df <- FCSex(Species_list=Cucurbita_splist,
                                      Occurrence_data=CucurbitaData,
                                      Raster_list=CucurbitaRasters,
                                      Buffer_distance=50000,
                                      Ecoregions_shp=ecoregions,
                                      Gap_MapEx=FALSE)


#Running all three In-situ gap analysis steps using FCSin function
FCSin_df <- FCSin(Species_list=Cucurbita_splist,
                                      Occurrence_data=CucurbitaData,
                                      Raster_list=CucurbitaRasters,
                                      Ecoregions_shp=ecoregions,
                                      Pro_areas=ProtectedAreas,
                                      Gap_MapIn=FALSE)

FCSc_mean_df <- FCSc_mean(FCSex_df = FCSex_df,FCSin_df = FCSin_df)


Aichi13_indicator_df  <- Aichi13_indicator(Species_list,FCSc_mean_df)

}
\references{
Ramirez-Villegas, J., Khoury, C., Jarvis, A., Debouck, D. G., & Guarino, L. (2010).
A Gap Analysis Methodology for Collecting Crop Genepools: A Case Study with Phaseolus Beans.
PLOS ONE, 5(10), e13497. Retrieved from https://doi.org/10.1371/journal.pone.0013497

Khoury, C. K., Amariles, D., Soto, J. S., Diaz, M. V., Sotelo, S., Sosa, C. C., … Jarvis, A. (2019).
Comprehensiveness of conservation of useful wild plants: An operational indicator for biodiversity
and sustainable development targets. Ecological Indicators. https://doi.org/10.1016/j.ecolind.2018.11.016
}
