% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkEcoregion.R
\name{checkEcoregion}
\alias{checkEcoregion}
\title{Check Ecoregion for compatability}
\usage{
checkEcoregion(eco, sdm, uniqueID)
}
\arguments{
\item{uniqueID}{}
}
\description{
Check Ecoregion for compatability
}
