% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkProtectedAreas.R
\name{checkProtectedAreas}
\alias{checkProtectedAreas}
\title{Quality check on protected areas dataset}
\usage{
checkProtectedAreas(protectedAreas, sdm)
}
\arguments{
\item{protectedAreas}{A terra rast object the contian spatial location of protected areas.}

\item{sdm}{a terra rast object that represented the expected distribution of the species}
}
\value{
protectedAreas : a terra rast object representing protected areas
}
\description{
Checks the class, values, crs, and cell size of the protected areas raster to ensure
these elements match those required by gap analysis functions.
}
\examples{

##Obtaining Raster_list

}
\references{
Khoury et al. (2019) Ecological Indicators 98:420-429. doi: 10.1016/j.ecolind.2018.11.016
Carver et al. (2021) GapAnalysis: an R package to calculate conservation indicators using spatial information
}
