% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkRasters.R
\name{checksdm}
\alias{checksdm}
\title{Quality check on sdm imagery}
\usage{
checksdm(sdm)
}
\arguments{
\item{sdm}{a terra rast object that represented the expected distribution of the species}
}
\value{
sdm : a terra rast object that is in the correct CRS
}
\description{
Evaluates the class, crs, and values are standardizes to what the following gap analysis functions are required.
}
\examples{
#' @examples
##Obtaining Raster_list
load("data/CucurbitaRasts.rda")

# convert the dataset for function
sdm <- terra::unwrap(CucurbitaRasts)$cordata
#Running checksdm
sdm <- checksdm(sdm = sdm)

}
\references{
Khoury et al. (2019) Ecological Indicators 98:420-429. doi: 10.1016/j.ecolind.2018.11.016
Carver et al. (2021) GapAnalysis: an R package to calculate conservation indicators using spatial information
}
