% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CucurbitaData}
\alias{CucurbitaData}
\title{Cucurbita occurrences dataset}
\format{
A data frame with 1184 rows and 4 variables:
\describe{
  \item{taxon}{character: Species name}
  \item{latitude}{numeric:Latitude in decimal format}
  \item{longitude}{numeric: Longitude in decimal format}
  \item{type}{character: Source of the record,germplasm (G) or herbarium (H)}
}
}
\source{
\url{https://dataverse.harvard.edu/dataverse/GapAnalysisR}
}
\usage{
CucurbitaData
}
\description{
This dataset is a subset of the original dataset for:
 C. cordata, C. digitata and C. palmata used in Khoury et al., 2020
}
\references{
Khoury, CK, Carver, D, Kates, HR, et al. Distributions, conservation status,
and abiotic stress tolerance potential of wild cucurbita (Cucurbita L.).
Plants, People, Planet. 2019; 00: 1– 15. https://doi.org/10.1002/ppp3.10085
}
\keyword{datasets}
