% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CucurbitaRasters}
\alias{CucurbitaRasters}
\title{Cucurbita species distribution models dataset}
\format{
raster files
}
\source{
\url{https://dataverse.harvard.edu/dataverse/GapAnalysisR}
}
\usage{
CucurbitaRasters
}
\description{
This dataset is a subset of species distribution models for:
 C. cordata, C. digitata and C. palmata used in Khoury et al., 2020
}
\references{
Khoury et al. (2019) Diversity and Distributions 26(2):209-225. doi: 10.1111/DDI.1300
}
\keyword{datasets}
