% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoregions}
\alias{ecoregions}
\title{Ecoregions vector}
\format{
SpatVector
}
\source{
\doi{10.7910/DVN/B8YOQL}
}
\usage{
ecoregions
}
\description{
This dataset is a subset of the Terrestrial Ecoregions of the World
 shapefile was made available by the world wildlife foundation
}
\keyword{datasets}
