% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ers_Insitu.R
\name{ers_Insitu}
\alias{ers_Insitu}
\title{Environmental representativeness score estimation (In situ conservation)}
\usage{
ers_Insitu(species_list, occurrenceData, raster_list, proArea, ecoReg)
}
\arguments{
\item{species_list}{An species list to calculate the ERSin metrics.}

\item{occurrenceData}{A data frame object with the species name, geographical coordinates, and type of records (G or H) for a given species}

\item{raster_list}{A list representing the species distribution models for the species list provided loaded in raster format. This list must match the same order of the species list.}

\item{proArea}{A raster file representing protected areas information. If proArea=NULL the funtion will use a protected area raster file
provided for your use after run preparing_Datasets()}

\item{ecoReg}{A shapefile representing ecoregions information with a field ECO_NUM representing ecoregions Ids. If ecoReg=NULL the funtion will use a shapefile
provided for your use after run preparing_Datasets()}
}
\value{
This function returns a data frame with two columns:

\tabular{lcc}{
species \tab Species name \cr
ERSin \tab ERSin value calculated\cr
}
}
\description{
This function performs an estimation of germplasm representativeness score for in-situ gap analysis (GRSin) using Khoury et al., (2019) methodology
This function uses a germplasm buffer raster file (e.g. CA50), a thresholded species distribution model, and a raster file of protected areas
 \deqn{ERSin = min(100,(Number of ecoregions of germplasm occurrences in protected areas/
Number of ecoregions of predicted Habitat within protected areas)*100)}
}
\examples{
##Obtaining occurrences from example
data(cucurbitaData)
##Obtaining species names from the data
speciesList <- unique(cucurbitaData$taxon)
##Obtaining raster_list
data(sdm_rasters)
##Obtaining protected areas raster
data(protectedArea)
##Obtaining ecoregions shapefile
data(ecoregions)

ERSin <- ers_Insitu(species_list = speciesList,
                   occurrenceData = cucurbitaData,
                   raster_list = sdm_rasters,
                   proArea= protectedArea,
                   ecoReg=ecoregions)

}
\references{
Ramirez-Villegas, J., Khoury, C., Jarvis, A., Debouck, D. G., & Guarino, L. (2010).
A Gap Analysis Methodology for Collecting Crop Genepools: A Case Study with Phaseolus Beans.
PLOS ONE, 5(10), e13497. Retrieved from https://doi.org/10.1371/journal.pone.0013497

Khoury, C. K., Amariles, D., Soto, J. S., Diaz, M. V., Sotelo, S., Sosa, C. C., … Jarvis, A. (2019).
Comprehensiveness of conservation of useful wild plants: An operational indicator for biodiversity
and sustainable development targets. Ecological Indicators. https://doi.org/10.1016/j.ecolind.2018.11.016
}
