% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERSin.R
\name{ERSin}
\alias{ERSin}
\title{Ecological representativeness score in situ}
\usage{
ERSin(
  taxon,
  sdm,
  occurrenceData,
  protectedAreas,
  ecoregions,
  idColumn,
  limitByPoints = FALSE
)
}
\arguments{
\item{taxon}{A character object that defines the name of the species as listed in the occurrence dataset}

\item{sdm}{a terra rast object that represented the expected distribution of the species}

\item{occurrenceData}{a data frame of values containing columns for the taxon, latitude, longitude, and type}

\item{protectedAreas}{A terra rast object the contian spatial location of protected areas.}

\item{ecoregions}{A terra vect object the contains spatial information on all ecoregions of interests}

\item{idColumn}{A character vector that notes what column within the ecoregions object should be used as a unique ID}

\item{limitByPoints}{A boolean parameter (TRUE/FALSE) to determine if you want to limit the ecoregions considered to those with observations present.
TRUE will exclude all ecoregions with no points within. FALSE will include all ecoregions.
This was implemented to prevent edge effects where pixels from the distribution extend into
neighboring ecoregions as a product of differences in raster/vector geometry rather than being predicted there directly.}
}
\value{
A list object containing
1. results : a data frames of values summarizing the results of the function
2. missingEcos : a terra vect object showing all the ecoregions within the distribution with no protected areas present
3. map : a leaflet object showing the spatial results of the function
}
\description{
The ERSin process provides an ecological measurement of the proportion of a species range
 that can be considered to be conserved in protected areas. The ERSin calculates the proportion of ecoregions
 encompassed within the range of the taxon located inside protected areas to the ecoregions encompassed
 within the total area of the distribution model, considering comprehensive conservation to have been accomplished
 only when every ecoregion potentially inhabited by a species is included within the distribution of the species
 located within a protected area.
}
\examples{
##Obtaining occurrences from example
data(CucurbitaData)
##Obtaining Raster_list
data(CucurbitaRasts)
##Obtaining protected areas raster
data(ProtectedAreas)
## ecoregion features
data(ecoregions)

# convert the dataset for function
taxon <- "Cucurbita_cordata"
sdm <- terra::unwrap(CucurbitaRasts)$cordata
protectedAreas <- terra::unwrap(ProtectedAreas)
ecoregions <- terra::vect(ecoregions)

#Running ERSin
ers_insitu <- ERSin(taxon = taxon,
                    sdm = sdm,
                    occurrenceData = CucurbitaData,
                    protectedAreas = protectedAreas,
                    ecoregions = ecoregions,
                    idColumn = "ECO_NAME",
                    limitByPoints = FALSE
                    )


}
\references{
Khoury et al. (2019) Ecological Indicators 98:420-429. \doi{10.1016/j.ecolind.2018.11.016}
Carver et al. (2021) GapAnalysis: an R package to calculate conservation indicators using spatial information
}
