% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExsituCompile.R
\name{ExsituCompile}
\alias{ExsituCompile}
\title{Ex-situ gap analysis calculation (Ex-situ conservation)}
\usage{
ExsituCompile(
  species_list,
  occurrenceData,
  raster_list,
  bufferDistance,
  ecoReg
)
}
\arguments{
\item{species_list}{A species list to calculate metrics.}

\item{occurrenceData}{A data frame object with the species name, geographical coordinates, and type of records (G or H) for a given species}

\item{raster_list}{A list representing the species distribution models for the species list provided loaded in raster format. This list must match the same order of the species list.}

\item{bufferDistance}{Geographical distance used to create circular buffers around germplasm. Default: 50000 that is 50 km around germplasm accessions (CA50)}

\item{ecoReg}{A shapefile representing ecoregions information with a field ECO_NUM representing ecoregions Ids. If ecoReg=NULL the funtion will use a shapefile
provided for your use after run preparing_Datasets()}
}
\value{
This function returns a data frame summarizing the ex-situ gap analysis scores:

\tabular{lcc}{
species \tab Species name \cr
SRSex \tab Ex-situ sample representativeness score \cr
GRSex \tab Ex-situ germplasm representativeness score \cr
ERSex \tab Ex-situ environmental representativeness score \cr
FCSex \tab Ex-situ final conservation score \cr
}
}
\description{
This function allows calculate all the three Ex situ gap analysis scores in one unique function returning a final conservation score summary table
}
\examples{
##Obtaining occurrences from example
data(CucurbitaData)
##Obtaining species names from the data
speciesList <- unique(CucurbitaData$taxon)
##Obtaining raster_list
data(CucurbitaRasters)
##Obtaining ecoregions shapefile
data(Ecoregions)

#Running all three Ex situ gap analysis steps using ExsituCompile function
exsituGapMetrics <- ExsituCompile(species_list=speciesList,
                                      occurrenceData=CucurbitaData,
                                      raster_list=CucurbitaRasters,
                                      bufferDistance=50000,
                                      ecoReg=Ecoregions)

}
\references{
Ramirez-Villegas, J., Khoury, C., Jarvis, A., Debouck, D. G., & Guarino, L. (2010).
A Gap Analysis Methodology for Collecting Crop Genepools: A Case Study with Phaseolus Beans.
PLOS ONE, 5(10), e13497. Retrieved from https://doi.org/10.1371/journal.pone.0013497

Khoury, C. K., Amariles, D., Soto, J. S., Diaz, M. V., Sotelo, S., Sosa, C. C., … Jarvis, A. (2019).
Comprehensiveness of conservation of useful wild plants: An operational indicator for biodiversity
and sustainable development targets. Ecological Indicators. https://doi.org/10.1016/j.ecolind.2018.11.016
}
