% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_combine.R
\name{fcs_Combine}
\alias{fcs_Combine}
\title{Combining Ex situ and In situ gap analysis results in one comprehensive conservation score (Summary Assessments)}
\usage{
fcs_Combine(fcsEx, fcsIn)
}
\arguments{
\item{fcsEx}{A data frame object result of the functions exsituGapAnalysis or fcs_exsitu}

\item{fcsIn}{A data frame object result of the functions insituGapAnalysis or fcs_insitu}
}
\value{
this function returns a data frame object with the following columns:

\tabular{lcc}{
 species \tab Species name \cr
 FCSex \tab Ex situ final conservation score \cr
 FCSin \tab In situ final conservation score \cr
 FCSc_min \tab Final conservation score (mininum value between FCSin and FCSex) \cr
 FCSc_max \tab Final conservation score (maximum value between FCSin and FCSex) \cr
 FCSc_mean \tab Final conservation score (average value between FCSin and FCSex) \cr
 FCSc_min_class \tab Final conservation category using  FCSc_min value \cr
 FCSc_max_class \tab Final conservation category using  FCSc_max value \cr
 FCSc_mean_class \tab Final conservation category using  FCSc_mean value \cr
}
}
\description{
This function concatenates ex situ conservation scores (SRSex, GRSex, ERSex,FCSex), and  in situ scores (SRSin, GRSin, ERSin,FCSin)
in one unique table and calculate the final conservation score for a species using Khoury et al., (2019) methodology.
}
\examples{
##Obtaining occurrences from example
data("cucurbitaData")
##Obtaining species names from the data
speciesList <- unique(cucurbitaData$taxon)
##Obtaining raster_list
data("sdm_rasters")
##Obtaining protected areas raster
data("protectedArea")
##Obtaining ecoregions shapefile
data("ecoregions")

#Running all three Ex situ gap analysis steps using insituGapAnalysis function
exsituGapMetrics <- exsitu_GapAnalysis(species_list=speciesList,
                                      occurrenceData=cucurbitaData,
                                      raster_list=sdm_rasters,
                                      bufferDistance=50000,
                                      ecoReg=ecoregions)


#Running all three In situ gap analysis steps using insituGapAnalysis function
insituGapMetrics <- insitu_GapAnalysis(species_list=speciesList,
                                       occurrenceData=cucurbitaData,
                                       raster_list=sdm_rasters,
                                       proArea=protectedArea,
                                       ecoReg=ecoregions)

fcsCombine <- fcs_Combine(fcsEx = exsituGapMetrics,fcsIn = insituGapMetrics)

}
\references{
Ramírez-Villegas, J., Khoury, C., Jarvis, A., Debouck, D. G., & Guarino, L. (2010).
A Gap Analysis Methodology for Collecting Crop Genepools: A Case Study with Phaseolus Beans.
PLOS ONE, 5(10), e13497. Retrieved from https://doi.org/10.1371/journal.pone.0013497

Khoury, C. K., Amariles, D., Soto, J. S., Diaz, M. V., Sotelo, S., Sosa, C. C., … Jarvis, A. (2019).
Comprehensiveness of conservation of useful wild plants: An operational indicator for biodiversity
and sustainable development targets. Ecological Indicators. https://doi.org/10.1016/j.ecolind.2018.11.016
}
