% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_Exsitu.R
\name{fcs_Exsitu}
\alias{fcs_Exsitu}
\title{Final ex situ conservation score estimation (Ex situ conservation)}
\usage{
fcs_Exsitu(srsDF, grsDF, ersDF)
}
\arguments{
\item{srsDF}{A dataframe object result of the srs_exsitu function}

\item{grsDF}{A dataframe object result of the grs_exsitu function}

\item{ersDF}{A dataframe object result of the ers_exsitu function}
}
\value{
This function returns a data frame with the follows information summarizing the ex situ gap analysis scores:

\tabular{lcc}{
species \tab Species name \cr
SRSex \tab Ex situ sample representativeness score \cr
GRSex \tab Ex situ germplasm representativeness score \cr
ERSex \tab Ex situ environmental representativeness score \cr
FCSex \tab Ex situ final conservation score \cr
}
}
\description{
This function concatenates the SRSex, GRSex, and ERSex values in an unique dataframe object to calculate a final priority score as
the average of the SRSex, GRSex, and ERSex values:

\deqn{FCSex = mean(SRSex,GRSex,ERSex)}
}
\examples{
##Obtaining occurrences from example
data("cucurbitaData")
##Obtaining species names from the data
speciesList <- unique(cucurbitaData$taxon)
##Obtaining raster_list
data("sdm_rasters")
##Obtaining ecoregions shapefile
data("ecoregions")

#Calculating SRSex value
SRSex <- srs_Exsitu(species_list = speciesList,
                    occurrenceData = cucurbitaData)

#Calculating GRSex value
GRSex <- grs_Exsitu(species_list = speciesList,
                    occurrenceData = cucurbitaData,
                    raster_list = sdm_rasters)

#Calculating ERSex value
ERSex <- ers_Exsitu(species_list = speciesList,
                    occurrenceData = cucurbitaData,
                    raster_list = sdm_rasters,
                    bufferDistance = 50000,
                    ecoReg=ecoregions
                    )

#Calculating final conservation for ex-situ gap analysis

fcsEx <- fcs_Exsitu(srsDF = SRSex,grsDF = GRSex,ersDF = ERSex)

}
\references{
Ramírez-Villegas, J., Khoury, C., Jarvis, A., Debouck, D. G., & Guarino, L. (2010).
A Gap Analysis Methodology for Collecting Crop Genepools: A Case Study with Phaseolus Beans.
PLOS ONE, 5(10), e13497. Retrieved from https://doi.org/10.1371/journal.pone.0013497

Khoury, C. K., Amariles, D., Soto, J. S., Diaz, M. V., Sotelo, S., Sosa, C. C., … Jarvis, A. (2019).
Comprehensiveness of conservation of useful wild plants: An operational indicator for biodiversity
and sustainable development targets. Ecological Indicators. https://doi.org/10.1016/j.ecolind.2018.11.016
}
