% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateGBuffers.R
\name{generateGBuffers}
\alias{generateGBuffers}
\title{Generate buffer of G type occurrences}
\usage{
generateGBuffers(taxon, occurrenceData, bufferDistM)
}
\arguments{
\item{taxon}{A character object that defines the name of the species as listed in the occurrence dataset}

\item{occurrenceData}{a data frame of values containing columns for the taxon, latitude, longitude, and type. Coordinates are assumed to be in the WGS84 (EPSG:4326) coordinate reference system.}

\item{bufferDistM}{Distance in meters. Used to set the size of the buffered objects.}
}
\value{
A list object containing
1. data : a terra vect object showing all the buffered areas around the G type occurrences
2. map : a leaflet object showing the spatial results of the function
}
\description{
Produces a terra vect object representing the area around the G type occurrences
}
\examples{
##Obtaining occurrences from example
data(CucurbitaData)

# convert the dataset for function
taxon <- "Cucurbita_cordata"
occurrenceData <- CucurbitaData

#Running generateGBuffers
gBuffer <- generateGBuffers(taxon = taxon,
                    occurrenceData = occurrenceData,
                    bufferDistM = 50000
                    )



}
\references{
Khoury et al. (2019) Ecological Indicators 98:420-429. \doi{10.1016/j.ecolind.2018.11.016}
Carver et al. (2021) GapAnalysis: an R package to calculate conservation indicators using spatial information
}
