% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsituCompile.R
\name{InsituCompile}
\alias{InsituCompile}
\title{In-situ gap analysis calculation (In-situ conservation)}
\usage{
InsituCompile(species_list, occurrenceData, raster_list, proArea, ecoReg)
}
\arguments{
\item{species_list}{An species list to calculate metrics.}

\item{occurrenceData}{A data frame object with the species name, geographical coordinates, and type of records (G or H) for a given species}

\item{raster_list}{A list representing the species distribution models for the species list provided loaded in raster format. This list must match the same order of the species list.}

\item{proArea}{A raster file representing protected areas information. If proArea=NULL the function will use a protected area raster file
provided for your use after run GetDatasets()}

\item{ecoReg}{A shapefile representing ecoregions information with a field ECO_NUM representing ecoregions Ids. If ecoReg=NULL the function will use a shapefile
provided for your use after run GetDatasets()}
}
\value{
This function returns a data frame summarizing the in-situ gap analysis scores:

\tabular{lcc}{
species \tab Species name \cr
SRSin \tab In-situ sample representativeness score \cr
GRSin \tab In-situ germplasm representativeness score \cr
ERSin \tab In-situ environmental representativeness score \cr
FCSin \tab In-situ final conservation score \cr
}
}
\description{
This function allows calculating all the three In-situ gap analysis scores in one unique function returning a final conservation score summary table
}
\examples{
##Obtaining occurrences from example
data(CucurbitaData)
##Obtaining species names from the data
speciesList <- unique(CucurbitaData$taxon)
##Obtaining raster_list
data(CucurbitaRasters)
CucurbitaRasters <- raster::unstack(CucurbitaRasters)
##Obtaining protected areas raster
data(ProtectedAreas)
##Obtaining ecoregions shapefile
data(ecoregions)

#Running all three In situ gap analysis steps using InsituCompile function
insituGapMetrics <- InsituCompile(species_list=speciesList,
                                       occurrenceData=CucurbitaData,
                                       raster_list=CucurbitaRasters,
                                       proArea=ProtectedAreas,
                                       ecoReg=ecoregions)

}
\references{
Ramirez-Villegas, J., Khoury, C., Jarvis, A., Debouck, D. G., & Guarino, L. (2010).
A Gap Analysis Methodology for Collecting Crop Genepools: A Case Study with Phaseolus Beans.
PLOS ONE, 5(10), e13497. Retrieved from https://doi.org/10.1371/journal.pone.0013497

Khoury, C. K., Amariles, D., Soto, J. S., Diaz, M. V., Sotelo, S., Sosa, C. C., … Jarvis, A. (2019).
Comprehensiveness of conservation of useful wild plants: An operational indicator for biodiversity
and sustainable development targets. Ecological Indicators. https://doi.org/10.1016/j.ecolind.2018.11.016
}
