% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OccurrenceCounts.R
\name{OccurrenceCounts}
\alias{OccurrenceCounts}
\title{Generating Species Count Data Frame}
\usage{
OccurrenceCounts(species, occurrenceData)
}
\arguments{
\item{species}{The name of a species}

\item{occurrenceData}{A data frame object with the species name, g
geographical coordinates, and type of records (G or H) for a given species}
}
\value{
This function returns a data frame object with the following columns:

\tabular{lcc}{
 species \tab Species name \cr
 totalRecords  \tab Total records available \cr
 hasLat  \tab Number of occurrences with latitude \cr
 hasLon  \tab Number of occurrences with longitude \cr
 totalUseful  \tab Number of occurrences with longitude \cr
 totalGRecords  \tab Number of occurrences with longitude \cr
 totalGUseful  \tab Number of occurrences with longitude \cr
 totalHRecords  \tab Number of occurrences with longitude \cr
 totalHUseful  \tab Number of occurrences with longitude \cr
 }
}
\description{
This function creates a summary file the summarizes
 the total number of germplasm and herbarium occurrences
}
\examples{
data(CucurbitaData)
##Obtaining species names from the data
speciesList <- unique(CucurbitaData$taxon)
countDF <- OccurrenceCounts(speciesList[[1]], CucurbitaData)

}
\references{
Ramirez-Villegas, J., Khoury, C., Jarvis, A., Debouck, D. G., & Guarino, L. (2010).
A Gap Analysis Methodology for Collecting Crop Genepools: A Case Study with Phaseolus Beans.
PLOS ONE, 5(10), e13497. Retrieved from https://doi.org/10.1371/journal.pone.0013497

Khoury, C. K., Amariles, D., Soto, J. S., Diaz, M. V., Sotelo, S., Sosa, C. C., … Jarvis, A. (2019).
Comprehensiveness of conservation of useful wild plants: An operational indicator for biodiversity
and sustainable development targets. Ecological Indicators. https://doi.org/10.1016/j.ecolind.2018.11.016
}
