% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing_Datasets.R
\name{preparing_Datasets}
\alias{preparing_Datasets}
\title{Preparing datasets to run gapAnalysisR functions}
\usage{
preparing_Datasets()
}
\value{
This function downloads data required to run a full gap analysis. It creates the folder data_preloaded into the gapAnalysisR.
Please run this function after install the package.
}
\description{
This function creates the data_preloaded folder and it downloads datasets from dataverse to allow run gapAnalysis package function.
}
\examples{
\dontrun{
preparing_Datasets()
}

}
\references{
Khoury, CK, Carver, D, Kates, HR, et al. Distributions, conservation status, and abiotic stress tolerance potential of wild cucurbits (Cucurbita L.).
Plants, People, Planet. 2019; 00: 1– 15. https://doi.org/10.1002/ppp3.10085
}
