% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ProtectedAreas}
\alias{ProtectedAreas}
\title{Protected areas dataset in raster format}
\format{
Raster file
}
\source{
\url{https://www.protectedplanet.net/}
}
\usage{
ProtectedAreas
}
\description{
This dataset is a raster version of the world protected areas dataset
 used in Khoury et al., (2019)
}
\references{
Khoury et al. (2019) Ecological Indicators 98:420-429. doi: 10.1016/j.ecolind.2018.11.016
}
\keyword{datasets}
