% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sdm_rasters}
\alias{sdm_rasters}
\title{Cucurbita species distribution models dataset}
\format{
raster files
}
\source{
\url{https://dataverse.harvard.edu/dataverse/GapAnalysisR}
}
\usage{
sdm_rasters
}
\description{
This dataset is a subset of thresholded species distribution models for:
 C. cordata, C. digitata and C. palmata used in Khoury et al., 2020
}
\references{
Khoury, CK, Carver, D, Kates, HR, et al. Distributions, conservation status,
 and abiotic stress tolerance potential of wild cucurbita (Cucurbita L.).
 Plants, People, Planet. 2019; 00: 1– 15. https://doi.org/10.1002/ppp3.10085
}
\keyword{datasets}
