% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRSin.R
\name{SRSin}
\alias{SRSin}
\title{Sampling representativeness score in situ}
\usage{
SRSin(taxon, sdm, occurrenceData, protectedAreas)
}
\arguments{
\item{taxon}{A character object that defines the name of the species as listed in the occurrence dataset}

\item{sdm}{a terra rast object that}

\item{occurrenceData}{a data frame of values containing columns for the taxon, latitude, longitude, and type. Coordinates are assumed to be in the WGS84 (EPSG:4326) coordinate reference system.}

\item{protectedAreas}{A terra rast object the contain spatial location of protected areas.}
}
\value{
A list object containing
1. results : a data frames of values summarizing the results of the function
2. points : a terra vect object showing all the points present within protected areas
3. map : a leaflet object showing the spatial results of the function
}
\description{
The SRSin process calculates the proportion of all occurrences of a taxon falling within
the distribution model that also fall within a protected area
}
\examples{
##Obtaining occurrences from example
data(CucurbitaData)
##Obtaining Raster_list
data(CucurbitaRasts)
##Obtaining protected areas raster
data(ProtectedAreas)

# convert the dataset for function
taxon <- "Cucurbita_cordata"
sdm <- terra::unwrap(CucurbitaRasts)$cordata
occurrenceData <- CucurbitaData
protectedAreas <- terra::unwrap(ProtectedAreas)
#Running SRSin
srs_insitu <- SRSin(taxon = taxon,
                    sdm = sdm,
                    occurrenceData = occurrenceData,
                    protectedAreas = protectedAreas
                    )

}
\references{
Khoury et al. (2019) Ecological Indicators 98:420-429. \doi{10.1016/j.ecolind.2018.11.016}
Carver et al. (2021) GapAnalysis: an R package to calculate conservation indicators using spatial information
}
