% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkEcoregion.R
\name{checkEcoregion}
\alias{checkEcoregion}
\title{Quality check of ecoregion dataset}
\usage{
checkEcoregion(ecoregion, sdm, idColumn)
}
\arguments{
\item{ecoregion}{A terra vect object the contains spatial information on all ecoregions of interests}

\item{sdm}{a terra rast object that represented the expected distribution of the species}

\item{idColumn}{A character vector that notes what column within the ecoregions object should be used as a unique ID}
}
\value{
ecoregions : A terra vect object the contains spatial information on all ecoregions of interests
}
\description{
Checks the class, crs, if the idColumn is a unique ID,
}
\examples{
##Obtaining Raster_list

}
\references{
Khoury et al. (2019) Ecological Indicators 98:420-429. doi: 10.1016/j.ecolind.2018.11.016
Carver et al. (2021) GapAnalysis: an R package to calculate conservation indicators using spatial information
}
