% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkEcoregion.R
\name{checkEcoregion}
\alias{checkEcoregion}
\title{Quality check of ecoregion dataset}
\usage{
checkEcoregion(ecoregion, sdm, idColumn)
}
\arguments{
\item{sdm}{a terra rast object that represented the expected distribution of the species}

\item{idColumn}{A character vector that notes what column within the ecoregions object should be used as a unique ID}

\item{ecoregions}{A terra vect object the contains spatial information on all ecoregions of interests}
}
\value{
ecoregions : A terra vect object the contains spatial information on all ecoregions of interests
}
\description{
Checks the class, crs, if the idColumn is a unique ID,
}
\examples{
##Obtaining Raster_list
load("data/CucurbitaRasts.rda")
## ecoregion features
load("data/ecoExample.rda")

# convert the dataset for function
sdm <- terra::unwrap(CucurbitaRasts)$cordata
ecoregions <- terra::vect(eco1)
#Running checkEcoregion
ecoregions <- checkEcoregion(ecoregion = ecoregions,
                    sdm = sdm,
                    idColumn = "ECO_NAME"
                    )


}
\references{
Khoury et al. (2019) Ecological Indicators 98:420-429. doi: 10.1016/j.ecolind.2018.11.016
Carver et al. (2021) GapAnalysis: an R package to calculate conservation indicators using spatial information
}
