% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkOccurrences.R
\name{checkOccurrences}
\alias{checkOccurrences}
\title{Quality check on occurrences data}
\usage{
checkOccurrences(csv, taxon, removeDuplicated = FALSE)
}
\arguments{
\item{csv}{A dataframe holding the occurrence data}

\item{taxon}{A character object that defines the name of the species as listed in the occurrence dataset}

\item{removeDuplicated}{: Binary parameter. TRUE == duplication values are remove. Set to FALSE as default}
}
\value{
A list object containing
1. data : a data frames of values of occurrence data in the required format
2. map : a leaflet object showing the spatial results of the function
}
\description{
Checks the column names, column data types, valid lat lon, and can optionally remove any duplicated lat lon records per species.
The cleaned and formated dataframe is returned as well as a map object show a quick reference of the points in space.
}
\examples{
# example code

##Obtaining occurrences from example


}
\references{
Khoury et al. (2019) Ecological Indicators 98:420-429. doi: 10.1016/j.ecolind.2018.11.016
Carver et al. (2021) GapAnalysis: an R package to calculate conservation indicators using spatial information
}
