% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CucurbitaData}
\alias{CucurbitaData}
\title{Cucurbita occurrences dataset}
\format{
A data frame with 1184 rows and 4 variables:
\describe{
  \item{species}{character: Species name}
  \item{latitude}{numeric:Latitude in decimal format}
  \item{longitude}{numeric: Longitude in decimal format}
  \item{type}{character: Source of the record,germplasm (G) or herbarium (H)}
}
}
\source{
\url{https://dataverse.harvard.edu/dataverse/GapAnalysis}
}
\usage{
CucurbitaData
}
\description{
This dataset is a subset of the original dataset for:
 C. cordata, C. digitata and C. palmata used in Khoury et al. (2019)
}
\references{
Khoury et al. (2019) Plants, People, Planet 2(3):269-283. doi: 10.1002/ppp3.10085.
}
\keyword{datasets}
