% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoregions}
\alias{ecoregions}
\title{Ecoregions vector}
\format{
SpatVector
}
\source{
\url{https://www.worldwildlife.org/publications/terrestrial-ecoregions-of-the-world}
}
\usage{
ecoregions
}
\description{
This dataset is a subset of the Terrestrial Ecoregions of the World
 shapefile made available by the world wildlife foundation
}
\keyword{datasets}
