% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoregions}
\alias{ecoregions}
\title{Ecoregions shapefile}
\format{
Shapefile
}
\source{
\url{http://maps.tnc.org/gis_data.html}
}
\usage{
ecoregions
}
\description{
This dataset is a subset of the Terrestrial Ecoregions shapefile made by the Nature Conservancy
}
\keyword{datasets}
