% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERSex.R
\name{ERSex}
\alias{ERSex}
\title{Ecological representativeness score ex situ}
\usage{
ERSex(
  taxon,
  sdm,
  occurrenceData,
  gBuffer,
  ecoregions,
  idColumn,
  limitByPoints = FALSE
)
}
\arguments{
\item{taxon}{A character object that defines the name of the species as listed in the occurrence dataset}

\item{sdm}{a terra rast object that represented the expected distribution of the species}

\item{occurrenceData}{a data frame of values containing columns for the taxon, latitude, longitude, and type. Coordinates are assumed to be in the WGS84 (EPSG:4326) coordinate reference system.}

\item{gBuffer}{A terra vect which encompases a specific buffer distance around all G points}

\item{ecoregions}{A terra vect object the contains spatial information on all ecoregions of interests}

\item{idColumn}{A character vector that notes what column within the ecoregions object should be used as a unique ID}

\item{limitByPoints}{A boolean parameter (TRUE/FALSE) to determine if you want to limit the ecoregions considered to those with observations present.
TRUE will exclude all ecoregions with no points within. FALSE will include all ecoregions.
This was implemented to prevent edge effects where pixels from the distribution extend into
neighboring ecoregions as a product of differences in raster/vector geometry rather than being predicted there directly.}
}
\value{
A list object containing
1. results : a data frames of values summarizing the results of the function
2. ecogaps : a terra vect object showing the ecoregions with no area within the g buffer objects
3. map : a leaflet object showing the spatial results of the function
}
\description{
The ERSex process provides an ecological measurement of the proportion of a species
 range that can be considered to be conserved in ex situ repositories. The ERSex calculates the
 proportion of terrestrial ecoregions (The Nature Conservancy Geospatial Conservation Atlas 2019)
 represented within the G buffered areas out of the total number of ecoregions occupied by the distribution model.
}
\examples{
##Obtaining occurrences from example
data(CucurbitaData)
##Obtaining Raster_list
data(CucurbitaRasts)
## ecoregion features
data(ecoregions)

# convert the dataset for function
taxon <- "Cucurbita_digitata"
sdm <- terra::unwrap(CucurbitaRasts)$digitata
ecoregions <- terra::vect(ecoregions)
#Running generateGBuffers
gBuffer <- generateGBuffers(taxon = taxon,
                    occurrenceData = CucurbitaData,
                    bufferDistM = 50000
                    )
#Running ERSex
ers_exsitu <- ERSex(taxon = taxon,
                    sdm = sdm,
                    occurrenceData = CucurbitaData,
                   gBuffer = gBuffer,
                   ecoregions = ecoregions,
                   idColumn = "ECO_NAME",
                   limitByPoints = FALSE
                   )


}
\references{
Khoury et al. (2019) Ecological Indicators 98:420-429. \doi{10.1016/j.ecolind.2018.11.016}
Carver et al. (2021) GapAnalysis: an R package to calculate conservation indicators using spatial information
}
