% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERSex.R
\name{ERSex}
\alias{ERSex}
\title{Environmental representativeness score estimation (Ex-situ conservation)}
\usage{
ERSex(
  Species_list,
  Occurrence_data,
  Raster_list,
  Buffer_distance = 50000,
  Ecoregions_shp = NULL
)
}
\arguments{
\item{Species_list}{An species list to calculate the ERSex metrics.}

\item{Occurrence_data}{A data frame object with the species name, geographical coordinates,
and type of records (G or H) for a given species}

\item{Raster_list}{A list representing the species distribution models for the species list provided
loaded in raster format. This list must match the same order of the species list.}

\item{Buffer_distance}{Geographical distance used to create circular buffers around germplasm.
Default: 50000 that is 50 km around germplasm accessions (CA50)}

\item{Ecoregions_shp}{A shapefile representing Ecoregionsions information with a field ECO_NUM representing Ecoregionsions Ids.
If Ecoregions=NULL the function will use a shapefile provided for your use after run GetDatasets()}
}
\value{
This function returns a data frame with two columns:

\tabular{lcc}{
species \tab Species name \cr
ERSex \tab ERSex value calculated\cr
}
}
\description{
This function performs an estimation of the environmental representativeness
 score for ex-situ gap analysis (ERSex) using Ramirez-Villegas et al., (2010) methodology.
ERSex is calculated as:
\deqn{ERSex = min(100,(Number of Ecoregionsions with 50km of G Occurrences /
 Number of Ecoregionsions Present within the Predict habitat)*100)}
}
\examples{
##Obtaining occurrences from example
data(CucurbitaData)
Cucurbita_splist <- unique(CucurbitaData$taxon)
## Obtaining rasterList object. ##
data(CucurbitaRasters)
CucurbitaRasters <- raster::unstack(CucurbitaRasters)
##Obtaining Ecoregionsions shapefile
data(ecoregions)

ERSex_df <- ERSex(Species_list = Cucurbita_splist,
                    Occurrence_data = CucurbitaData,
                    Raster_list = CucurbitaRasters,
                    Buffer_distance = 50000,
                    Ecoregions_shp=ecoregions)

}
\references{
Ramirez-Villegas, J., Khoury, C., Jarvis, A., Debouck, D. G., & Guarino, L. (2010).
A Gap Analysis Methodology for Collecting Crop Genepools: A Case Study with Phaseolus Beans.
PLOS ONE, 5(10), e13497. Retrieved from https://doi.org/10.1371/journal.pone.0013497

Khoury, C. K., Amariles, D., Soto, J. S., Diaz, M. V., Sotelo, S., Sosa, C. C., … Jarvis, A. (2019).
Comprehensiveness of conservation of useful wild plants: An operational indicator for biodiversity
and sustainable development targets. Ecological Indicators. https://doi.org/10.1016/j.ecolind.2018.11.016
}
