% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERSex.R
\name{ERSex}
\alias{ERSex}
\title{Environmental representativeness score estimation (Ex-situ conservation)}
\usage{
ERSex(species_list, occurrenceData, raster_list, bufferDistance, ecoReg)
}
\arguments{
\item{species_list}{An species list to calculate the ERSex metrics.}

\item{occurrenceData}{A data frame object with the species name, geographical coordinates, and type of records (G or H) for a given species}

\item{raster_list}{A list representing the species distribution models for the species list provided loaded in raster format. This list must match the same order of the species list.}

\item{bufferDistance}{Geographical distance used to create circular buffers around germplasm. Default: 50000 that is 50 km around germplasm accessions (CA50)}

\item{ecoReg}{A shapefile representing ecoregions information with a field ECO_NUM representing ecoregions Ids. If ecoReg=NULL the function will use a shapefile
provided for your use after run GetDatasets()}
}
\value{
This function returns a data frame with two columns:

\tabular{lcc}{
species \tab Species name \cr
ERSex \tab ERSex value calculated\cr
}
}
\description{
This function performs an estimation of the environmental representativeness score for ex-situ gap analysis (ERSex) using Ramirez-Villegas et al., (2010) methodology.
ERSex is calculated as:
\deqn{ERSex = min(100,(Number of ecoregions with 50km of G Occurrences / Number of Ecoregions Present within the Predict habitat)*100)}
}
\examples{
##Obtaining occurrences from example
data(CucurbitaData)
speciesList <- unique(CucurbitaData$taxon)
## Obtaining rasterList object. ##
data(CucurbitaRasters)
##Obtaining ecoregions shapefile
data(Ecoregions)

ERSex_df <- ERSex(species_list = speciesList,
                    occurrenceData = CucurbitaData,
                    raster_list = CucurbitaRasters,
                    bufferDistance = 50000,
                    ecoReg=Ecoregions)

}
\references{
Ramirez-Villegas, J., Khoury, C., Jarvis, A., Debouck, D. G., & Guarino, L. (2010).
A Gap Analysis Methodology for Collecting Crop Genepools: A Case Study with Phaseolus Beans.
PLOS ONE, 5(10), e13497. Retrieved from https://doi.org/10.1371/journal.pone.0013497

Khoury, C. K., Amariles, D., Soto, J. S., Diaz, M. V., Sotelo, S., Sosa, C. C., … Jarvis, A. (2019).
Comprehensiveness of conservation of useful wild plants: An operational indicator for biodiversity
and sustainable development targets. Ecological Indicators. https://doi.org/10.1016/j.ecolind.2018.11.016
}
