% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERSin.R
\name{ERSin}
\alias{ERSin}
\title{Ecological representativeness score in situ}
\usage{
ERSin(
  Species_list,
  Occurrence_data,
  Raster_list,
  Pro_areas = NULL,
  Ecoregions_shp = NULL,
  Gap_Map = NULL
)
}
\arguments{
\item{Species_list}{A species list to calculate the ERSin metric}

\item{Occurrence_data}{A data frame object with the species name, geographical coordinates,
and type of records (G or H) for a given species}

\item{Raster_list}{A list representing the species distribution models for the species list provided
loaded in raster format. This list must match the same order of the species list.}

\item{Pro_areas}{A raster file representing protected areas information.
If Pro_areas=NULL the funtion will use a protected area raster file provided for your use after run GetDatasets()}

\item{Ecoregions_shp}{A shapefile representing Ecoregions_shp information with a field ECO_NUM representing Ecoregions_shp Ids.}

\item{Gap_Map}{Default=FALSE, This option will calculate gap maps for each species analyzed and will retun a list
with two slots ERSin and gap_maps}
}
\value{
This function returns a data frame with two columns:

\tabular{lcc}{
species \tab Species name \cr
ERSin \tab ERSin value calculated\cr
}
}
\description{
The ERSin process provides an ecological measurement of the proportion of a species range
 that can be considered to be conserved in protected areas. The ERSin calculates the proportion of ecoregions
 encompassed within the range of the taxon located inside protected areas to the ecoregions encompassed
 within the total area of the distribution model, considering comprehensive conservation to have been accomplished
 only when every ecoregion potentially inhabited by a species is included within the distribution of the species
 located within a protected area.
 This function uses a thresholded species distribution model, an ecoregions file, and a raster file of protected areas
}
\examples{
##Obtaining occurrences from example
data(CucurbitaData)
##Obtaining species names from the data
Cucurbita_splist <- unique(CucurbitaData$taxon)
##Obtaining Raster_list
data(CucurbitaRasters)
CucurbitaRasters <- raster::unstack(CucurbitaRasters)
##Obtaining protected areas raster
data(ProtectedAreas)
##Obtaining ecoregions shapefile
data(ecoregions)

ERSin_df <- ERSin(Species_list = Cucurbita_splist,
                   Occurrence_data = CucurbitaData,
                   Raster_list = CucurbitaRasters,
                   Pro_areas= ProtectedAreas,
                   Ecoregions_shp=ecoregions)

}
\references{
Khoury et al. (2019) Ecological Indicators 98:420-429. doi: 10.1016/j.ecolind.2018.11.016
}
