% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCSc_mean.R
\name{FCSc_mean}
\alias{FCSc_mean}
\title{Final Conservation Score measure}
\usage{
FCSc_mean(taxon, fcsin, fcsex)
}
\arguments{
\item{taxon}{A character object that defines the name of the species as listed in the occurrence dataset}

\item{fcsin}{A data frame containing summary results from the fcsin function}

\item{fcsex}{A data frame containing summary results from the fcsex function}
}
\value{
data_comb : a data frame which aggreates final result summaries
}
\description{
Compiles all tabular data from the individual metrics and generate the final results
}
\examples{
##Obtaining occurrences from example
data(CucurbitaData)
##Obtaining Raster_list
data(CucurbitaRasts)
##Obtaining protected areas raster
data(ProtectedAreas)
## ecoregion features
data(ecoregions)

# convert the dataset for function
taxon <- "Cucurbita_cordata"
sdm <- terra::unwrap(CucurbitaRasts)$cordata
occurrenceData <- CucurbitaData
protectedAreas <- terra::unwrap(ProtectedAreas)
ecoregions <- terra::vect(ecoregions)


# generate exsitu conservation summaries
srs_exsitu <- SRSex(taxon = taxon,
                    occurrence_Data  = CucurbitaData
                    )

gBuffer <- generateGBuffers(taxon = taxon,
                    occurrenceData = occurrenceData,
                    bufferDistM = 50000
                    )#'

grs_exsitu <- GRSex(taxon = taxon,
                    sdm = sdm,
                    gBuffer = gBuffer
                    )

ers_exsitu <- ERSex(taxon = taxon,
                    sdm = sdm,
                    occurrence_Data = occurrenceData,
                    gBuffer = gBuffer,
                    ecoregions = ecoregions,
                    idColumn = "ECO_NAME"
                    )

#Running fcsin
fcs_exsitu <- FCSex(taxon = taxon,
                    srsex = srs_exsitu,
                    grsex = grs_exsitu,
                    ersex = ers_exsitu
                    )


# generate insitu conservation summaries
srs_insitu <- SRSin(taxon = taxon,
                    sdm = sdm,
                    occurrenceData = CucurbitaData,
                    protectedAreas = protectedAreas
                    )

grs_insitu <- GRSin(taxon = taxon,
                    sdm = sdm,
                    protectedAreas = protectedAreas
                    )

ers_insitu <- ERSin(taxon = taxon,
                    sdm = sdm,
                    occurrenceData = occurrenceData,
                    protectedAreas = protectedAreas,
                    ecoregions = ecoregions,
                    idColumn = "ECO_NAME"
                    )

#Running fcsin
fcs_insitu <- FCSin(taxon = taxon,
                    srsin = srs_insitu,
                    grsin = grs_insitu,
                    ersin = ers_insitu
                    )

fsc_combine <- FCSc_mean(taxon = taxon,
                         fcsin = fcs_insitu,
                         fcsex = fcs_exsitu)

}
\references{
Khoury et al. (2019) Ecological Indicators 98:420-429. doi: 10.1016/j.ecolind.2018.11.016
Carver et al. (2021) GapAnalysis: an R package to calculate conservation indicators using spatial information
}
