% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCSex.R
\name{FCSex}
\alias{FCSex}
\title{Final conservation score ex situ}
\usage{
FCSex(taxon, srsex, grsex, ersex)
}
\arguments{
\item{taxon}{}

\item{ersex}{}
}
\description{
This function calculates the average of the three ex situ conservation metrics
  returning a final conservation score summary table. It also assigns conservation priority categories
}
\references{
Khoury et al. (2019) Ecological Indicators 98:420-429. doi: 10.1016/j.ecolind.2018.11.016
Carver et al. (2021) GapAnalysis: an R package to calculate conservation indicators using spatial information
}
