% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCSex.R
\name{FCSex}
\alias{FCSex}
\title{Final conservation score ex situ}
\usage{
FCSex(
  Species_list,
  Occurrence_data,
  Raster_list,
  Buffer_distance = 50000,
  Ecoregions_shp = NULL
)
}
\arguments{
\item{Species_list}{A species list to calculate metrics.}

\item{Occurrence_data}{A data frame object with the species name, geographical coordinates,
and type of records (G or H) for a given species}

\item{Raster_list}{A list representing the species distribution models for the species list provided
loaded in raster format. This list must match the same order of the species list.}

\item{Buffer_distance}{Geographical distance used to create circular buffers around germplasm.
Default: 50000 (50 km) around germplasm accessions (CA50)}

\item{Ecoregions_shp}{A shapefile representing ecoregions information with a field ECO_NUM representing ecoregions Ids.
If ecoReg=NULL the funtion will use a shapefile provided for your use after run GetDatasets()}
}
\value{
This function returns a list with gap maps if Gap_MapEx=TRUE. Otherwise, it returns a data frame
summarizing the ex-situ gap analysis scores:

\tabular{lcc}{
species \tab Species name \cr
SRSex \tab Sampling representativeness score ex situ \cr
GRSex \tab Geographical representativeness score ex situ \cr
ERSex \tab Ecological representativeness score ex situ \cr
FCSex \tab Final conservation score ex situ  \cr
}
}
\description{
This function calculates the average of the three ex situ conservation metrics
  returning a final conservation score summary table.
}
\examples{
##Obtaining occurrences from example
data(CucurbitaData)
##Obtaining species names from the data
Cucurbita_splist <- unique(CucurbitaData$taxon)
##Obtaining raster_list
data(CucurbitaRasters)
CucurbitaRasters <- raster::unstack(CucurbitaRasters)
##Obtaining ecoregions shapefile
data(ecoregions)

#Running all three Ex-situ gap analysis steps using a unique function
FCSex_df <- FCSex(Species_list=Cucurbita_splist,
                                      Occurrence_data=CucurbitaData,
                                      Raster_list=CucurbitaRasters,
                                      Buffer_distance=50000,
                                      Ecoregions_shp=ecoregions)

}
\references{
Khoury et al. (2019) Ecological Indicators 98:420-429. doi: 10.1016/j.ecolind.2018.11.016
}
