% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCSex.R
\name{FCSex}
\alias{FCSex}
\title{Final conservation score ex situ}
\usage{
FCSex(taxon, srsex, grsex, ersex)
}
\arguments{
\item{taxon}{A character object that defines the name of the species as listed in the occurrence dataset}

\item{srsex}{A dataframe contain the results from the srsex function}

\item{grsex}{A dataframe contain the results from the grsex function}

\item{ersex}{A dataframe contain the results from the ersex function}
}
\value{
out_df : a data frames of values summarizing the results of the function
}
\description{
This function calculates the average of the three ex situ conservation metrics
  returning a final conservation score summary table. It also assigns conservation priority categories
}
\examples{
##Obtaining occurrences from example
data(CucurbitaData)
##Obtaining Raster_list
data(CucurbitaRasts)
##Obtaining protected areas raster
data(ProtectedAreas)
## ecoregion features
data(ecoregions)

# convert the dataset for function
taxon <- "Cucurbita_cordata"
sdm <- terra::unwrap(CucurbitaRasts)$cordata
occurrenceData <- CucurbitaData
protectedAreas <- terra::unwrap(ProtectedAreas)
ecoregions <- terra::unwrap(ecoregions)
# generate exsitu conservation summaries
srs_exsitu <- SRSex(taxon = taxon,
                    occurrence_Data = CucurbitaData
                    )

gBuffer <- generateGBuffers(taxon = taxon,
                    occurrenceData = occurrenceData,
                    bufferDistM = 50000
                    )#'

grs_exsitu <- GRSex(taxon = taxon,
                    sdm = sdm,
                    gBuffer = gBuffer
                    )

ers_exsitu <- ERSex(taxon = taxon,
                    sdm = sdm,
                    occurrence_Data = occurrenceData,
                    gBuffer = gBuffer,
                    ecoregions = ecoregions,
                    idColumn = "ECO_NAME"
                    )

#Running fcsex
fcs_exsitu <- FCSex(taxon = taxon,
                    srsex = srs_exsitu,
                    grsex = grs_exsitu,
                    ersex = ers_exsitu)



}
\references{
Khoury et al. (2019) Ecological Indicators 98:420-429. doi: 10.1016/j.ecolind.2018.11.016
Carver et al. (2021) GapAnalysis: an R package to calculate conservation indicators using spatial information
}
