% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCSex.R
\name{FCSex}
\alias{FCSex}
\title{Final ex situ conservation score estimation (Ex-situ conservation)}
\usage{
FCSex(
  Species_list,
  Occurrence_data,
  Raster_list,
  Buffer_distance,
  Ecoregions_shp,
  Gap_MapEx
)
}
\arguments{
\item{Species_list}{A species list to calculate metrics.}

\item{Occurrence_data}{A data frame object with the species name, geographical coordinates,
and type of records (G or H) for a given species}

\item{Gap_MapEx}{Default=FALSE, This option will calculate gap maps for each species analyzed and will retun a list
with two slots FCSex and gap_maps}

\item{raster_list}{A list representing the species distribution models for the species list provided
loaded in raster format. This list must match the same order of the species list.}

\item{bufferDistance}{Geographical distance used to create circular buffers around germplasm.
Default: 50000 that is 50 km around germplasm accessions (CA50)}

\item{ecoReg}{A shapefile representing ecoregions information with a field ECO_NUM representing ecoregions Ids.
If ecoReg=NULL the funtion will use a shapefile provided for your use after run GetDatasets()}
}
\value{
This function returns a list with gap maps if Gap_MapEx=TRUE. Otherwise, it returns a data frame
summarizing the ex-situ gap analysis scores:

\tabular{lcc}{
species \tab Species name \cr
SRSex \tab Ex-situ sample representativeness score \cr
GRSex \tab Ex-situ germplasm representativeness score \cr
ERSex \tab Ex-situ environmental representativeness score \cr
FCSex \tab Ex-situ final conservation score \cr
}
}
\description{
This function allows calculate all the three Ex-situ gap analysis scores
  in one unique function returning a final conservation score summary table.
  The final priority score is calculated as the average of the SRSex, GRSex, and ERSex values:

\deqn{FCSex = mean(SRSex,GRSex,ERSex)}
}
\examples{
##Obtaining occurrences from example
data(CucurbitaData)
##Obtaining species names from the data
Cucurbita_splist <- unique(CucurbitaData$taxon)
##Obtaining raster_list
data(CucurbitaRasters)
CucurbitaRasters <- raster::unstack(CucurbitaRasters)
##Obtaining ecoregions shapefile
data(ecoregions)

#Running all three Ex-situ gap analysis steps using a unique function
FCSex_df <- FCSex(Species_list=Cucurbita_splist,
                                      Occurrence_data=CucurbitaData,
                                      Raster_list=CucurbitaRasters,
                                      Buffer_distance=50000,
                                      Ecoregions_shp=ecoregions,
                                      Gap_MapEx=FALSE)

}
\references{
Ramirez-Villegas, J., Khoury, C., Jarvis, A., Debouck, D. G., & Guarino, L. (2010).
A Gap Analysis Methodology for Collecting Crop Genepools: A Case Study with Phaseolus Beans.
PLOS ONE, 5(10), e13497. Retrieved from https://doi.org/10.1371/journal.pone.0013497

Khoury, C. K., Amariles, D., Soto, J. S., Diaz, M. V., Sotelo, S., Sosa, C. C., … Jarvis, A. (2019).
Comprehensiveness of conservation of useful wild plants: An operational indicator for biodiversity
and sustainable development targets. Ecological Indicators. https://doi.org/10.1016/j.ecolind.2018.11.016
}
