% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCSin.R
\name{FCSin}
\alias{FCSin}
\title{Final conservation score in situ}
\usage{
FCSin(taxon, srsin, grsin, ersin, noModel = FALSE)
}
\arguments{
\item{taxon}{A character object that defines the name of the species as listed in the occurrence dataset}

\item{srsin}{A dataframe contain the results from the srsin function}

\item{grsin}{A dataframe contain the results from the grsin function}

\item{ersin}{A dataframe contain the results from the ersin function}

\item{noModel}{A boolean parameter (TRUE/FALSE) to determine if there is a species distribution model. Default is FALSE.
When TRUE, GRS and ERS cannot be calculated and are assigned 0; the final score is the mean of the three metrics,
following Khoury et al. (2019).}
}
\value{
out_df : a data frames of values summarizing the results of the function
}
\description{
This function calculates the average of the three in situ conservation metrics and
assigns a priority category based on the results
}
\examples{
##Obtaining occurrences from example
data(CucurbitaData)
##Obtaining Raster_list
data(CucurbitaRasts)
##Obtaining protected areas raster
data(ProtectedAreas)
## ecoregion features
data(ecoregions)

# convert the dataset for function
taxon <- "Cucurbita_cordata"
sdm <- terra::unwrap(CucurbitaRasts)$cordata
occurrenceData <- CucurbitaData
protectedAreas <- terra::unwrap(ProtectedAreas)
ecoregions <- terra::vect(ecoregions)

# generate insitu conservation summaries
srs_insitu <- SRSin(taxon = taxon,
                    sdm = sdm,
                    occurrenceData = occurrenceData,
                    protectedAreas = protectedAreas
                    )

grs_insitu <- GRSin(taxon = taxon,
                    sdm = sdm,
                    protectedAreas = protectedAreas
                    )

ers_insitu <- ERSin(taxon = taxon,
                    sdm = sdm,
                    occurrenceData = occurrenceData,
                    protectedAreas = protectedAreas,
                    ecoregions = ecoregions,
                    idColumn = "ECO_NAME"
                    )

#Running fcsin
FCSin <- FCSin(taxon = taxon,
                    srsin = srs_insitu,
                    grsin = grs_insitu,
                    ersin = ers_insitu
                    )



}
\references{
Khoury et al. (2019) Ecological Indicators 98:420-429. \doi{10.1016/j.ecolind.2018.11.016}
Carver et al. (2021) GapAnalysis: an R package to calculate conservation indicators using spatial information
}
