% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCSin.R
\name{FCSin}
\alias{FCSin}
\title{Final in-situ conservation score estimation (In-situ conservation)}
\usage{
FCSin(srsDF, grsDF, ersDF)
}
\arguments{
\item{srsDF}{A dataframe object result of the SRSin function}

\item{grsDF}{A dataframe object result of the GRSin function}

\item{ersDF}{A dataframe object result of the ERSin function}
}
\value{
This function returns a data frame with the follows information
 summarizing the in-situ gap analysis scores:

\tabular{lcc}{
species \tab Species name \cr
SRSin \tab In-situ sample representativeness score \cr
GRSin \tab In-situ germplasm representativeness score \cr
ERSin \tab In-situ environmental representativeness score \cr
FCSin \tab In-situ final conservation score \cr
}
}
\description{
This function concatenates the SRSin, GRSin, and ERSin values
in an unique dataframe object to calculate a final priority score as
the average of the SRSin, GRSin, and ERSin values:

\deqn{FCSin = mean(SRSin,GRSin,ERSin)}
}
\examples{
##Obtaining occurrences from example
data(CucurbitaData)
##Obtaining species names from the data
speciesList <- unique(CucurbitaData$taxon)
##Obtaining raster_list
data(CucurbitaRasters)
CucurbitaRasters <- raster::unstack(CucurbitaRasters)
##Obtaining protected areas raster
data(ProtectedAreas)
##Obtaining ecoregions shapefile
data(ecoregions)
#Calculating SRSin value
SRSin_df <- SRSin(species_list = speciesList,
                    occurrenceData = CucurbitaData,
                    raster_list = CucurbitaRasters,
                    proArea=ProtectedAreas)

#Calculating GRSin value
GRSin_df <- GRSin(species_list = speciesList,
                    occurrenceData = CucurbitaData,
                    raster_list = CucurbitaRasters,
                    proArea=ProtectedAreas)

#Calculating ERSin value
ERSin_df <- ERSin(species_list = speciesList,
                    occurrenceData = CucurbitaData,
                    raster_list = CucurbitaRasters,
                    proArea=ProtectedAreas,
                    ecoReg=ecoregions)

#Calculating final conservation for ex-situ gap analysis

FCSin_df <- FCSin(srsDF = SRSin_df, grsDF = GRSin_df, ersDF = ERSin_df)

}
\references{
Ramirez-Villegas, J., Khoury, C., Jarvis, A., Debouck, D. G., & Guarino, L. (2010).
A Gap Analysis Methodology for Collecting Crop Genepools: A Case Study with Phaseolus Beans.
PLOS ONE, 5(10), e13497. Retrieved from https://doi.org/10.1371/journal.pone.0013497

Khoury, C. K., Amariles, D., Soto, J. S., Diaz, M. V., Sotelo, S., Sosa, C. C., … Jarvis, A. (2019).
Comprehensiveness of conservation of useful wild plants: An operational indicator for biodiversity
and sustainable development targets. Ecological Indicators. https://doi.org/10.1016/j.ecolind.2018.11.016
}
