% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCSin.R
\name{FCSin}
\alias{FCSin}
\title{Final in-situ conservation score estimation (In-situ conservation)}
\usage{
FCSin(
  Species_list,
  Occurrence_data,
  Raster_list,
  Ecoregions_shp,
  Pro_areas,
  Gap_MapIn
)
}
\arguments{
\item{Species_list}{A species list to calculate metrics.}

\item{Occurrence_data}{A data frame object with the species name, geographical coordinates,
and type of records (G or H) for a given species}

\item{Raster_list}{A list representing the species distribution models for the species list provided
loaded in raster format. This list must match the same order of the species list.}

\item{Ecoregions_shp}{A shapefile representing Ecoregions_shpions information with a field ECO_NUM representing Ecoregions_shpions Ids.
If Ecoregions_shp=NULL the funtion will use a shapefile provided for your use after run GetDatasets()}

\item{Pro_areas}{A raster file representing protected areas information.
If Pro_areas=NULL the funtion will use a protected area raster file provided for your use after run GetDatasets()}

\item{Gap_MapIn}{Default=FALSE, This option will calculate gap maps for each species analyzed and will retun a list
with two slots FCSin and gap_maps}
}
\value{
This function returns a list with gap maps if Gap_MapIn=TRUE. Otherwise, it returns a data frame
 summarizing the in-situ gap analysis scores:

\tabular{lcc}{
species \tab Species name \cr
SRSin \tab In-situ sample representativeness score \cr
GRSin \tab In-situ germplasm representativeness score \cr
ERSin \tab In-situ environmental representativeness score \cr
FCSin \tab In-situ final conservation score \cr
}
}
\description{
TThis function allows calculate all the three In-situ gap analysis scores
in an unique dataframe object to calculate a final priority score as
the average of the SRSin, GRSin, and ERSin values:

\deqn{FCSin = mean(SRSin,GRSin,ERSin)}
}
\examples{
##Obtaining occurrences from example
data(CucurbitaData)
##Obtaining species names from the data
Cucurbita_splist <- unique(CucurbitaData$taxon)
##Obtaining Raster_list
data(CucurbitaRasters)
CucurbitaRasters <- raster::unstack(CucurbitaRasters)
##Obtaining protected areas raster
data(ProtectedAreas)
##Obtaining Ecoregions_shpions shapefile
data(ecoregions)
#Running all three In-situ gap analysis steps using a unique function
FCSin_df <- FCSin(Species_list=Cucurbita_splist,
                                      Occurrence_data=CucurbitaData,
                                      Raster_list=CucurbitaRasters,
                                      Ecoregions_shp=ecoregions,
                                      Pro_areas=ProtectedAreas,
                                      Gap_MapIn=FALSE)

}
\references{
Ramirez-Villegas, J., Khoury, C., Jarvis, A., Debouck, D. G., & Guarino, L. (2010).
A Gap Analysis Methodology for Collecting Crop Genepools: A Case Study with Phaseolus Beans.
PLOS ONE, 5(10), e13497. Retrieved from https://doi.org/10.1371/journal.pone.0013497

Khoury, C. K., Amariles, D., Soto, J. S., Diaz, M. V., Sotelo, S., Sosa, C. C., … Jarvis, A. (2019).
Comprehensiveness of conservation of useful wild plants: An operational indicator for biodiversity
and sustainable development targets. Ecological Indicators. https://doi.org/10.1016/j.ecolind.2018.11.016
}
