% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCSin.R
\name{FCSin}
\alias{FCSin}
\title{Final conservation score in situ}
\usage{
FCSin(
  Species_list,
  Occurrence_data,
  Raster_list,
  Ecoregions_shp = NULL,
  Pro_areas = NULL
)
}
\arguments{
\item{Species_list}{A species list to calculate metrics.}

\item{Occurrence_data}{A data frame object with the species name, geographical coordinates,
and type of records (G or H) for a given species}

\item{Raster_list}{A list representing the species distribution models for the species list provided
loaded in raster format. This list must match the same order of the species list.}

\item{Ecoregions_shp}{A shapefile representing Ecoregions_shp information with a field ECO_NUM
representing Ecoregions_shp Ids.}

\item{Pro_areas}{A raster file representing protected areas information.If Pro_areas=NULL the funtion will use
a protected area raster file provided for your use
after run GetDatasets()}
}
\value{
This function returns a list with gap maps if Gap_MapIn=TRUE. Otherwise,
it returns a data frame summarizing the in situ gap analysis scores:
\tabular{lcc}{
species \tab Species name \cr
SRSin \tab Sampling representativeness score in situ  \cr
GRSin \tab Geographical representativeness score in situ \cr
ERSin \tab Ecological representativeness score in situ \cr
FCSin \tab Final conservation score in situ  \cr
}
}
\description{
This function calculates the average of the three in situ conservation metrics
}
\examples{
##Obtaining occurrences from example
data(CucurbitaData)
##Obtaining species names from the data
Cucurbita_splist <- unique(CucurbitaData$taxon)
##Obtaining Raster_list
data(CucurbitaRasters)
CucurbitaRasters <- raster::unstack(CucurbitaRasters)
##Obtaining protected areas raster
data(ProtectedAreas)
##Obtaining Ecoregions_shpions shapefile
data(ecoregions)
#Running all three In-situ gap analysis steps using a unique function
FCSin_df <- FCSin(Species_list=Cucurbita_splist,
                                      Occurrence_data=CucurbitaData,
                                      Raster_list=CucurbitaRasters,
                                      Ecoregions_shp=ecoregions,
                                      Pro_areas=ProtectedAreas)

}
\references{
Khoury et al. (2019) Diversity and Distributions 26(2):209-225. doi: 10.1111/DDI.13008
}
