% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GapAnalysis-package.R
\docType{package}
\name{GapAnalysis-package}
\alias{GapAnalysis-package}
\alias{GapAnalysis}
\title{GapAnalysis: An R package to calculate conservation indicators using spatial information}
\description{
GapAnalysis is a an R package used to calculate ex situ and in situ conservation indicators
}
\details{
\tabular{ll}{
Package: \tab GapAnalysis\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2020-04-8\cr
License: \tab GPL-3\cr
}
}
\author{
\strong{CIAT Decision and Policy Analysis research area}\cr
email: \email{J.R.Villegas@cgiar.org}
}
\keyword{package}
