% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateEcoSelection.R
\name{generateEcoSelection}
\alias{generateEcoSelection}
\title{Title}
\usage{
generateEcoSelection(taxon, ococcurrence_Data, ecoregions, idColumn)
}
\arguments{
\item{idColumn}{}
}
\description{
Title
}
