% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDatasets.R
\name{GetDatasets}
\alias{GetDatasets}
\title{Preparing datasets to run GapAnalysis functions}
\usage{
GetDatasets()
}
\value{
This function downloads data required to run a full gap analysis.
 It creates the folder data_preloaded into the GapAnalysis.
Please run this function after install the package.
}
\description{
This function creates the data_preloaded folder.
 It downloads datasets from dataverse to allow run gapAnalysis package function.
}
\examples{
\dontrun{
GetDatasets()
}

}
\references{
Khoury, CK, Carver, D, Kates, HR, et al. Distributions, conservation status, and abiotic stress tolerance potential of wild cucurbits (Cucurbita L.).
Plants, People, Planet. 2019; 00: 1– 15. https://doi.org/10.1002/ppp3.10085
}
