% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCSc_mean.R, R/checkEcoregion.R,
%   R/checkOccurrences.R, R/checkProtectedAreas.R, R/checkRasters.R,
%   R/downloadData.R
\name{GetDataSets}
\alias{GetDataSets}
\alias{FCSc_mean}
\alias{checkEcoregion}
\alias{checkOccurrences}
\alias{checkProtectAreas}
\alias{checkRaster}
\alias{GetDatasets}
\title{Download datasets from Dataverse}
\usage{
FCSc_mean(taxon, fcsin, fcsex)

checkEcoregion(eco, sdm, uniqueID)

checkOccurrences(csv, taxon, removeDuplicated = FALSE)

checkProtectAreas(proArea, sdm)

checkRaster(raster)

GetDatasets()
}
\arguments{
\item{eco}{}

\item{sdm}{}

\item{uniqueID}{}

\item{csv}{}

\item{removeDuplicated}{}

\item{proArea}{}
}
\description{
A short description...

A short description...

A short description...

A short description...

A short description...

A short description...
}
\references{
Khoury et al. (2019) Ecological Indicators 98:420-429. doi: 10.1016/j.ecolind.2018.11.016
Carver et al. (2021) GapAnalysis: an R package to calculate conservation indicators using spatial information

Khoury et al. (2019) Ecological Indicators 98:420-429. doi: 10.1016/j.ecolind.2018.11.016
Carver et al. (2021) GapAnalysis: an R package to calculate conservation indicators using spatial information

Khoury et al. (2019) Ecological Indicators 98:420-429. doi: 10.1016/j.ecolind.2018.11.016
Carver et al. (2021) GapAnalysis: an R package to calculate conservation indicators using spatial information

Khoury et al. (2019) Ecological Indicators 98:420-429. doi: 10.1016/j.ecolind.2018.11.016
Carver et al. (2021) GapAnalysis: an R package to calculate conservation indicators using spatial information

Khoury et al. (2019) Ecological Indicators 98:420-429. doi: 10.1016/j.ecolind.2018.11.016
Carver et al. (2021) GapAnalysis: an R package to calculate conservation indicators using spatial information

Khoury et al. (2019) Ecological Indicators 98:420-429. doi: 10.1016/j.ecolind.2018.11.016
Carver et al. (2021) GapAnalysis: an R package to calculate conservation indicators using spatial information
}
