% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDatasets.R
\name{getDatasets}
\alias{getDatasets}
\title{Download datasets from the harvard dataverse repo}
\usage{
getDatasets()
}
\description{
Ecoregions and protected area data base are stored on a harvard dataverse repository. This functions check to see if
those datasets have been download and will download them if not present.
}
\examples{
getDatasets()

}
\references{
Khoury et al. (2019) Ecological Indicators 98:420-429. doi: 10.1016/j.ecolind.2018.11.016
Carver et al. (2021) GapAnalysis: an R package to calculate conservation indicators using spatial information
}
