% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDatasets.R
\name{GetDatasets}
\alias{GetDatasets}
\title{Preparing datasets to run GapAnalysis functions}
\usage{
GetDatasets()
}
\value{
This function downloads data required to run a full gap analysis.
 It creates the folder data_preloaded into the GapAnalysis.
Please run this function after installing the package.
}
\description{
This function creates the data_preloaded folder.
 It downloads datasets from dataverse to allow the gapAnalysis package to function.
}
\examples{
\dontrun{
GetDatasets()
}

}
\references{
Khoury et al. (2019) Data in Brief 22:90-97. doi: 10.1016/j.dib.2018.11.125.
Khoury et al. (2019) Diversity and Distributions 26(2):209-225. doi: 10.1111/DDI.1300
}
