% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRSex.R
\name{GRSex}
\alias{GRSex}
\title{Geographical representativeness score estimation (Ex-situ conservation)}
\usage{
GRSex(occurrenceData, species_list, raster_list, bufferDistance)
}
\arguments{
\item{occurrenceData}{A data frame object with the species name, geographical coordinates,
and type of records (G or H) for a given species}

\item{species_list}{An species list to calculate the GRSex metrics.}

\item{raster_list}{A list representing the species distribution models for the species list provided
loaded in raster format. This list must match the same order of the species list.}

\item{bufferDistance}{Geographical distance used to create circular buffers around germplasm.
Default: 50000 that is 50 km around germplasm accessions (CA50)}
}
\value{
This function returns a data frame with two columns:

\tabular{lcc}{
species \tab Species name \cr
GRSex \tab GRSex value calculated\cr
}
}
\description{
This function performs an estimation of the geographical representativeness score
 for ex-situ gap analysis (GRSex) using Ramirez-Villegas et al., (2010) methodology.
 GRS ex-situ score is calculated as:

\deqn{GRSex = min(100,(Masked Area of Buffered G Occurrences / Total Area of Predicted Habitat)*100)}
}
\examples{
##Obtaining occurrences from example
data(CucurbitaData)
speciesList <- unique(CucurbitaData$taxon)
## Obtaining rasterList objet. ##
data(CucurbitaRasters)
CucurbitaRasters <- raster::unstack(CucurbitaRasters)
#Calculating GRSex value
GRSex_df <- GRSex(species_list = speciesList,
                    occurrenceData = CucurbitaData,
                    raster_list = CucurbitaRasters,
                    bufferDistance = 50000)

}
\references{
Ramirez-Villegas, J., Khoury, C., Jarvis, A., Debouck, D. G., & Guarino, L. (2010).
A Gap Analysis Methodology for Collecting Crop Genepools: A Case Study with Phaseolus Beans.
PLOS ONE, 5(10), e13497. Retrieved from https://doi.org/10.1371/journal.pone.0013497

Khoury, C. K., Amariles, D., Soto, J. S., Diaz, M. V., Sotelo, S., Sosa, C. C., … Jarvis, A. (2019).
Comprehensiveness of conservation of useful wild plants: An operational indicator for biodiversity
and sustainable development targets. Ecological Indicators. https://doi.org/10.1016/j.ecolind.2018.11.016


Hijmans, R.J. and Spooner, D.M. (2001). Geographic distribution of wild potato species.
Am. J. Bot., 88: 2101-2112. doi:10.2307/3558435
}
