% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRSin.R
\name{GRSin}
\alias{GRSin}
\title{Geographical representativeness score estimation (In-situ conservation)}
\usage{
GRSin(Species_list, Occurrence_data, Raster_list, Pro_areas)
}
\arguments{
\item{Species_list}{An species list to calculate the GRSin metrics.}

\item{Occurrence_data}{A data frame object with the species name, geographical coordinates, and type of records (G or H) for a given species}

\item{Raster_list}{A list representing the species distribution models for the species list provided loaded in raster format. This list must match the same order of the species list.}

\item{Pro_areas}{A raster file representing protected areas information. If Pro_areas=NULL the function will use a protected area raster file
provided for your use after run GetDatasets()}
}
\value{
This function returns a data frame with two columns:

\tabular{lcc}{
species \tab Species name \cr
GRSin \tab GRSin value calculated\cr
}
}
\description{
This function performs an estimation of germplasm representativeness score for in-situ gap analysis (GRSin) using Khoury et al., (2019) methodology
#'  GRSIn is calculated as:
 \deqn{GRSin = min(100,(Predicted Habitat within protected areas/ Total Area of Predicted Habitat)*100)}
}
\examples{
##Obtaining occurrences from example
data(CucurbitaData)
##Obtaining species names from the data
Cucurbita_splist <- unique(CucurbitaData$taxon)
##Obtaining Raster_list
data(CucurbitaRasters)
CucurbitaRasters <- raster::unstack(CucurbitaRasters)
##Obtaining protected areas raster
data(ProtectedAreas)

GRSin_df <- GRSin(Species_list = Cucurbita_splist,
                    Occurrence_data = CucurbitaData,
                    Raster_list = CucurbitaRasters,
                    Pro_areas=ProtectedAreas)


}
\references{
Ramirez-Villegas, J., Khoury, C., Jarvis, A., Debouck, D. G., & Guarino, L. (2010).
A Gap Analysis Methodology for Collecting Crop Genepools: A Case Study with Phaseolus Beans.
PLOS ONE, 5(10), e13497. Retrieved from https://doi.org/10.1371/journal.pone.0013497

Khoury, C. K., Amariles, D., Soto, J. S., Diaz, M. V., Sotelo, S., Sosa, C. C., … Jarvis, A. (2019).
Comprehensiveness of conservation of useful wild plants: An operational indicator for biodiversity
and sustainable development targets. Ecological Indicators. https://doi.org/10.1016/j.ecolind.2018.11.016
}
