% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{indicator}
\alias{indicator}
\title{Using Ex-situ and In-situ gap analysis results to calculate the conservation indicator value}
\usage{
indicator(FCSc_mean_df)
}
\arguments{
\item{FCSc_mean_df}{A data frame object result of the function FCSc_mean}
}
\value{
this function returns a data frame object with the following columns:

\tabular{lcc}{
 opt \tab Final conservation score option \cr
 count_HP \tab Count of species in high priority status for collecting \cr
 count_MP \tab Count of species in medium priority status for collecting  \cr
 count_LP \tab Count of species in low priority status for collecting  \cr
 count_SC \tab Count of species sufficiently collected  \cr
 count_LP_SC Count of species sufficiently collected or in low  priority status for collecting   \tab \cr
 proportion_HP Proportion of species in high priority status for collecting\tab \cr
 proportion_MP \tab Proportion of species in medium priority status for collecting   \cr
 proportion_LP \tab Proportion of species in low priority status for collecting \cr
 proportion_SC \  Proportion of species sufficiently collected  \cr
 proportion_CP_SC \tab Proportion of species sufficiently collected or in low  priority status for collecting (Indicator) \cr
}
}
\description{
This function uses a data.frame resulting from the function FCSc_mean and
 computes a conservation progress indicator.
}
\examples{
##Obtaining occurrences from example
data(CucurbitaData)
##Obtaining species names from the data
Cucurbita_splist <- unique(CucurbitaData$taxon)
##Obtaining raster_list
data(CucurbitaRasters)
CucurbitaRasters <- raster::unstack(CucurbitaRasters)
##Obtaining protected areas raster
data(ProtectedAreas)
##Obtaining ecoregions shapefile
data(ecoregions)

#Running all three Ex-situ gap analysis steps using FCSex function
FCSex_df <- FCSex(Species_list=Cucurbita_splist,
                  Occurrence_data=CucurbitaData,
                  Raster_list=CucurbitaRasters,
                  Buffer_distance=50000,
                  Ecoregions_shp=ecoregions)


#Running all three In-situ gap analysis steps using FCSin function
FCSin_df <- FCSin(Species_list=Cucurbita_splist,
                  Occurrence_data=CucurbitaData,
                  Raster_list=CucurbitaRasters,
                  Ecoregions_shp=ecoregions,
                  Pro_areas=ProtectedAreas,
                  Gap_MapIn=FALSE)

FCSc_mean_df <- FCSc_mean(FCSex_df = FCSex_df,FCSin_df = FCSin_df)


indicator_df  <- indicator(FCSc_mean_df)

}
\references{
Khoury, C. K., Amariles, D., Soto, J. S., Diaz, M. V., Sotelo, S., Sosa, C. C., … Jarvis, A. (2019).
Comprehensiveness of conservation of useful wild plants: An operational indicator for biodiversity
and sustainable development targets. Ecological Indicators. https://doi.org/10.1016/j.ecolind.2018.11.016
}
