% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRSin.R
\name{SRSin}
\alias{SRSin}
\title{##Obtaining occurrences from example
load("data/CucurbitaData.rda")
##Obtaining species names from the data
taxon <- CucurbitaData$species[1]
##Obtaining Raster_list
load("data/CucurbitaRasts.rda")
##Obtaining protected areas raster
load("data/protectAreasRast.rda")
#' #Running SRSin}
\usage{
SRSin(taxon, sdm, occurrence_Data, protectedAreas)
}
\description{
##Obtaining occurrences from example
load("data/CucurbitaData.rda")
##Obtaining species names from the data
taxon <- CucurbitaData$species[1]
##Obtaining Raster_list
load("data/CucurbitaRasts.rda")
##Obtaining protected areas raster
load("data/protectAreasRast.rda")
#' #Running SRSin
}
