% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRSin.R
\name{SRSin}
\alias{SRSin}
\title{Sampling representativeness score in situ}
\usage{
SRSin(Species_list, Occurrence_data, Raster_list, Pro_areas = NULL)
}
\arguments{
\item{Species_list}{A species list to calculate the SRSin metric.}

\item{Occurrence_data}{A data frame object with the species name, geographical coordinates,
and type of records (G or H) for a given species}

\item{Raster_list}{A list representing the species distribution models for the species list
provided loaded in raster format. This list must match the same order of the species list.}

\item{Pro_areas}{A raster file representing protected areas information.
If Pro_areas=NULL the function will use a protected area raster file
provided for your use after run GetDatasets()}
}
\value{
This function returns a data frame with two columns:

\tabular{lcc}{
species \tab Species name \cr
SRSin \tab SRSin value calculated\cr
}
}
\description{
The SRSin process calculates the proportion of all occurrences of a taxon within the distribution model
that fall within a protected area
}
\examples{
##Obtaining occurrences from example
data(CucurbitaData)
##Obtaining species names from the data
Cucurbita_splist <- unique(CucurbitaData$taxon)
##Obtaining Raster_list
data(CucurbitaRasters)
CucurbitaRasters <- raster::unstack(CucurbitaRasters)
##Obtaining protected areas raster
data(ProtectedAreas)
SRSin_df <- SRSin(Species_list = Cucurbita_splist,
                    Occurrence_data = CucurbitaData,
                    Raster_list=CucurbitaRasters,
                    Pro_areas=ProtectedAreas)

}
\references{
Khoury et al. (2019) Diversity and Distributions 26(2):209-225. doi: 10.1111/DDI.13008.
}
