% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_HTML.R
\name{summary_HTML}
\alias{summary_HTML}
\title{Creating a summary HTML document for each taxon}
\usage{
summary_HTML(
  species_list,
  occurrenceData,
  raster_List,
  proArea,
  bufferDistance,
  insituSummary,
  exsituSummary,
  fcsSummary,
  outputFolder,
  writeRasters
)
}
\arguments{
\item{species_list}{A species list to calculate metrics.}

\item{occurrenceData}{A data frame object with the species name, geographical coordinates,
and type of records (G or H) for a given species}

\item{raster_List}{A list representing the species distribution models for the species list provided
loaded in raster format. This list must match the same order of the species list.}

\item{proArea}{A raster file representing protected areas information.
If proArea=NULL the funtion will use a protected area raster file provided for your use after run GetDatasets()}

\item{bufferDistance}{Geographical distance used to create circular buffers around germplasm.
Default: 50000 that is 50 km around germplasm accessions (CA50)}

\item{insituSummary}{A data frame object result of the FCSin}

\item{exsituSummary}{A data frame object result of the FCSex}

\item{fcsSummary}{A data frame object result of the function FCSc_mean}

\item{outputFolder}{A path to save the HTML file resulting of this function}

\item{writeRasters}{Boolean field (default=F) to indicate if raster files should be saved}
}
\value{
This function returns a data frame file saved at a specified folder
}
\description{
Calls the summaryHTML rmd file information and displays
 the quantitative and spatial results content.
}
\examples{
\dontrun{
##Obtaining occurrences from example
data(CucurbitaData)
##Obtaining species names from the data
Cucurbita_splist <- unique(CucurbitaData$taxon)
##Obtaining raster_list
data(CucurbitaRasters)
CucurbitaRasters <- raster::unstack(CucurbitaRasters)
##Obtaining protected areas raster
data(ProtectedAreas)
##Obtaining ecoregions shapefile
data(ecoregions)

#Running all three ex situ gap analysis steps
FCSex_df <- FCSex(species_list=speciesList,
                                      occurrenceData=CucurbitaData,
                                      raster_list=CucurbitaRasters,
                                      bufferDistance=50000,
                                      ecoReg=ecoregions)


#Running all three in situ gap analysis steps
FCSin_df <- FCSin(species_list=speciesList,
                                       occurrenceData=CucurbitaData,
                                       raster_list=CucurbitaRasters,
                                       proArea=ProtectedAreas,
                                       ecoReg=ecoregions)

fcsCombine <- FCSc_mean(fcsEx = FCSex_df,fcsIn = FCSin_df)

summaryHTML_file <- summary_HTML(species_list=Cucurbita_splist,
                                 occurrenceData = CucurbitaData,
                                 raster_List=CucurbitaRasters,
                                 proArea=ProtectedArea,
                                 bufferDistance=50000,
                                 insituSummary=FCSin_df,
                                 exsituSummary=FCSex_df,
                                 fcsSummary=fcsCombine,
                                 outputFolder=".",
                                 writeRasters=F)
}

}
\references{
Khoury et al. (2019) Diversity and Distributions 26(2):209-225. doi: 10.1111/DDI.13008
}
